New: Extracted scrolling and resizing logic from Gallery
